Move function coercion to its own module
